requirements.txt
openai-whisper
flask
ipython
ffmpeg-python
gunicorn